refactor:CatAnimationのAnimateとUnhappyOrHungryWalkinganimateを分離
